change to u64 because serde_json uses it as the number type in json

+++ src/author.rs
@@ -9,26 +9,26 @@ use crate::user::User;
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub(crate) struct Author {
     alias: MaybeInit<Box<Author>>,
-    alias_id: Option<u32>,
+    alias_id: Option<u64>,
     // todo!
     alternate_names: Value,
     bio: Option<String>,
-    books_count: u32,
+    books_count: u64,
     born_date: Option<Date>,
-    born_year: Option<u32>,
+    born_year: Option<u64>,
     cached_image: Image,
     canonical: MaybeInit<Box<Author>>,
-    canonical_id: Option<u32>,
+    canonical_id: Option<u64>,
     contributions: Vec<Contribution>,
     creator: Option<User>,
     death_date:	Option<Date>,
-    death_year: Option<u32>,
+    death_year: Option<u64>,
     gender_id: Option<Gender>,
-    id: u32,
+    id: u64,
     // todo!
     identifiers: Value,
     image: Option<Image>,
-    image_id: Option<u32>,
+    image_id: Option<u64>,
     is_bipoc: Option<bool>,
     is_lgbtq: Option<bool>,
     // todo!
@@ -41,6 +41,6 @@ pub(crate) struct Author {
     slug: Option<String>,
     state: RecordState,
     title: Option<String>,
-    user_id: Option<u32>,
-    users_count: u32,
+    user_id: Option<u64>,
+    users_count: u64,
 }
