+++ src/publisher.rs
@@ -4,18 +4,18 @@ use crate::edition::Edition;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub(crate) struct Publisher {
-    canonical_id: Option<u32>,
+    canonical_id: Option<u64>,
     created_at: PlainDateTime,
     editions: Vec<Edition>,
-    editions_count: u32,
-    id: u32,
+    editions_count: u64,
+    id: u64,
     locked: bool,
     name: Option<String>,
     object_type: String,
-    parent_id: Option<u32>,
+    parent_id: Option<u64>,
     parent_publisher: Option<Publisher>,
     slug: String,
     state: String,
     updated_at: PlainDateTime,
-    user_id: Option<u32>,
+    user_id: Option<u64>,
 }
