+++ src/tagging.rs
@@ -3,13 +3,13 @@ use serde_json::Value;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub(crate) struct Tagging {
-    count: u32,
-    id: u32,
+    count: u64,
+    id: u64,
     slug: String,
     tag: String,
     // todo!
     tag_category: Value,
-    tag_category_id: u32,
+    tag_category_id: u64,
     // todo!
     taggings: Vec<Value>,
 }
